Split yaml utils from file_utils.py
