+++ mlflow/store/artifact/models_artifact_repo.py
@@ -19,7 +19,6 @@ from mlflow.store.artifact.utils.models import (
     get_model_name_and_version,
     is_using_databricks_registry,
 )
-from mlflow.utils.file_utils import write_yaml
 from mlflow.utils.uri import (
     add_databricks_profile_info_to_artifact_uri,
     get_databricks_profile_uri_from_artifact_uri,
@@ -27,6 +26,7 @@ from mlflow.utils.uri import (
     is_models_uri,
     is_oss_unity_catalog_uri,
 )
+from mlflow.utils.yaml_utils import write_yaml
 
 REGISTERED_MODEL_META_FILE_NAME = "registered_model_meta"
