+++ tests/projects/test_virtualenv_projects.py
@@ -5,8 +5,8 @@ import pytest
 
 import mlflow
 from mlflow.exceptions import MlflowException
-from mlflow.utils.file_utils import read_yaml, write_yaml
 from mlflow.utils.virtualenv import _create_virtualenv
+from mlflow.utils.yaml_utils import read_yaml, write_yaml
 
 from tests.projects.utils import (
     TEST_VIRTUALENV_CONDA_PROJECT_DIR,
