+++ tests/store/tracking/test_file_store.py
@@ -44,12 +44,7 @@ from mlflow.store.tracking import SEARCH_MAX_RESULTS_DEFAULT
 from mlflow.store.tracking.file_store import FileStore
 from mlflow.tracing.constant import TraceMetadataKey, TraceTagKey
 from mlflow.tracking._tracking_service.utils import _use_tracking_uri
-from mlflow.utils.file_utils import (
-    TempDir,
-    path_to_local_file_uri,
-    read_yaml,
-    write_yaml,
-)
+from mlflow.utils.file_utils import TempDir, path_to_local_file_uri
 from mlflow.utils.mlflow_tags import (
     MLFLOW_DATASET_CONTEXT,
     MLFLOW_LOGGED_MODELS,
@@ -60,6 +55,7 @@ from mlflow.utils.os import is_windows
 from mlflow.utils.time import get_current_time_millis
 from mlflow.utils.uri import append_to_uri_path
 from mlflow.utils.validation import MAX_EXPERIMENT_NAME_LENGTH
+from mlflow.utils.yaml_utils import read_yaml, write_yaml
 
 from tests.helper_functions import random_int, random_str, safe_edit_yaml
